requirements.txt
fastapi
uvicorn
openai
langchain
pinecone
pydantic
tqdm
python-dotenv
pymongo
python-jose
python-multipart